pyproject.toml
[project]
name = "segmentation"
version = "0.1.0"
description = "Advanced image segmentation for interior spaces using YOLO, SAM 2, Mask2Former, and Gemini VLM"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ultralytics>=8.3.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "sam2>=1.0.0",
    "click>=8.0.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "transformers>=4.30.0",
    "timm>=0.9.0",
    "google-generativeai>=0.3.0",
    "scikit-learn>=1.0.0",
]
